+++ libs/interfaces/src/auth/admin_login.dto.ts
@@ -47,36 +47,3 @@ export class AdminLoginResponseDto {
   @IsString()
   token!: string;
 }
-
-export class AdminRegRequestDto extends BaseRequestDTO {
-  @ApiProperty({
-    description: '이메일',
-  })
-  @IsEmail()
-  email!: string;
-
-  @ApiProperty({
-    description: '패스워드',
-  })
-  @IsString()
-  password!: string;
-
-  @ApiProperty({
-    description: '이름',
-  })
-  @IsString()
-  name!: string;
-}
-export class AdminRegResponseDto {
-  @ApiProperty({
-    description: '이메일',
-  })
-  @IsEmail()
-  email!: string;
-
-  @ApiProperty({
-    description: '이름',
-  })
-  @IsString()
-  name!: string;
-}
